render to-do to the user interface
todos have been rendered to the user interface. Each todo icon on is a different colour. I had an issue while rendering the todos. Todo 1 was rendered 2 times and there was 2 renders coming from the 2nd render. I fixed this by clearing element before a re-render. The innerHTML equal to an empty string.